+++ src/main/java/zd/zero/waifu/motivator/plugin/personality/PersonalityCore.kt
@@ -1,0 +1,10 @@
+package zd.zero.waifu.motivator.plugin.personality
+
+import zd.zero.waifu.motivator.plugin.motivation.event.MotivationEvent
+
+interface PersonalityCore {
+    fun processMotivationEvent(
+            motivationEvent: MotivationEvent,
+            emotionalState: EmotionalState
+    )
+}
